рефактринг главной
