+++ src/pages/main-page/main-page.tsx
@@ -1,20 +1,16 @@
 import cn from 'classnames';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
 import CitiesFilter from '../../components/cities-filter/cities-filter';
 import Header from '../../components/header/header';
-import SortForm from '../../components/sort-form/sort-form';
 import NullOfferList from '../../components/null-offer-list/null-offer-list';
-import Map from '../../components/map/map';
-import MainOfferList from '../../components/proxy/main-offer-list';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { loadOfferList } from '../../store/action';
+import MainOffers from '../../components/main-offers/main-offers';
 
 export default function MainPage(): JSX.Element {
 
   const dispatch = useAppDispatch();
-  const [selectedOfferId, setSelectedOffer] = useState('');
-  const handleOfferSelect = (id: string) => setSelectedOffer(id);
   const currentCity = useAppSelector((state) => state.city);
   const offersData = useAppSelector((state) => state.offerList);
   const currentOfferList = offersData.filter((item) => item.city.name === currentCity?.name);
@@ -45,18 +41,9 @@ export default function MainPage(): JSX.Element {
             {'cities__places-container--empty': isOffersListEmpty},
             'container')}
           >
-            { !isOffersListEmpty ? (
-              <>
-                <section className="cities__places places">
-                  <h2 className="visually-hidden">Places</h2>
-                  <b className="places__found">{currentOfferList.length} places to stay in {currentCity?.name}</b>
-                  <SortForm />
-                  <MainOfferList offersList={currentOfferList} onOfferSelect={handleOfferSelect} />
-                </section>
-                <div className="cities__right-section">
-                  <Map city={currentCity} mapPoints={currentOfferList} selectedOfferId={selectedOfferId} className='cities__map' />
-                </div>
-              </>) : (<NullOfferList />)}
+            {!isOffersListEmpty ? (
+              <MainOffers currentCity={currentCity} currentOfferList={currentOfferList}/>
+            ) : (<NullOfferList />)}
           </div>
         </div>
       </main>
